backend/requirements.txt
Django>=5.0
djangorestframework
djangorestframework-simplejwt
django-cors-headers
python-decouple
psycopg2-binary
Pillow
supabase
django-filter
langchain
langchain-google-genai
langchain-community
langchain-text-splitters
langdetect
chromadb
sentence-transformers
google-generativeai
stripe
